--- rate-limiter-board/src/main/kotlin/com/yellowsunn/ratelimiterboard/const/HttpRequestConst.kt
@@ -1,5 +1,0 @@
-package com.yellowsunn.ratelimiterboard.const
-
-const val X_FORWARDED_FOR = "X-Forwarded-For"
-
-const val IP_PATTERN = "^\\d{1,3}.\\d{1,3}.\\d{1,3}.\\d{1,3}"
