--- rate-limiter-board/src/main/kotlin/com/yellowsunn/ratelimiterboard/exception/RateLimitException.kt
@@ -1,6 +1,0 @@
-package com.yellowsunn.ratelimiterboard.exception
-
-class RateLimitException(
-    key: String = "",
-    val burstCapacity: Long,
-) : IllegalStateException("Requests are limited due to too many requests. key=$key")
