+++ rate-limiter/build.gradle.kts
@@ -12,8 +12,8 @@ dependencies {
     implementation("org.springframework.boot:spring-boot-starter-data-redis")
     implementation("org.redisson:redisson-spring-boot-starter:${Versions.redisson}")
     implementation("com.fasterxml.jackson.module:jackson-module-kotlin")
-    implementation("com.github.yellowsunn.ratelimits:ratelimits-redis:1.0.1")
-    implementation("com.github.yellowsunn.ratelimits:ratelimits-core:1.0.1")
+    implementation("com.github.yellowsunn.ratelimits:ratelimits-redis:91c25f065a")
+    implementation("com.github.yellowsunn.ratelimits:ratelimits-core:91c25f065a")
     implementation("es.moki.ratelimitj:ratelimitj-redis:0.6.0")
     implementation("io.jsonwebtoken:jjwt-api:${Versions.jjwt}")
     runtimeOnly("io.jsonwebtoken:jjwt-impl:${Versions.jjwt}")
