+++ rate-limiter/src/main/kotlin/com/yellowsunn/ratelimiter/RateLimiterApplication.kt
@@ -1,11 +1,11 @@
-package com.yellowsunn.ratelimiterboard
+package com.yellowsunn.ratelimiter
 
 import org.springframework.boot.autoconfigure.SpringBootApplication
 import org.springframework.boot.runApplication
 
 @SpringBootApplication
-class RateLimiterBoardApplication
+class RateLimiterApplication
 
 fun main(args: Array<String>) {
-    runApplication<RateLimiterBoardApplication>(*args)
+    runApplication<RateLimiterApplication>(*args)
 }
