+++ rate-limiter/src/main/kotlin/com/yellowsunn/ratelimiter/config/WebConfiguration.kt
@@ -1,6 +1,6 @@
-package com.yellowsunn.ratelimiterboard.config
+package com.yellowsunn.ratelimiter.config
 
-import com.yellowsunn.ratelimiterboard.interceptor.RateLimitInterceptor
+import com.yellowsunn.ratelimiter.interceptor.RateLimitInterceptor
 import org.springframework.context.annotation.Configuration
 import org.springframework.web.servlet.config.annotation.InterceptorRegistry
 import org.springframework.web.servlet.config.annotation.WebMvcConfigurer
