+++ rate-limiter/src/main/kotlin/com/yellowsunn/ratelimiter/const/HttpRequestConst.kt
@@ -1,0 +1,3 @@
+package com.yellowsunn.ratelimiter.const
+
+const val X_FORWARDED_FOR = "X-Forwarded-For"
