+++ settings.gradle.kts
@@ -3,5 +3,5 @@ rootProject.name = "learning-projects"
 include(
     "kafka-graceful-shutdown",
     "stock-concurrency",
-    "rate-limiter-board",
+    "rate-limiter",
 )
